+++ frontend/app/src/screens/layout/breadcrumb-navigation/style.ts
@@ -1,2 +1,4 @@
 export const breadcrumbItemStyle =
   "w-auto h-auto border-none hover:bg-gray-100 py-1 px-2 rounded-md";
+
+export const breadcrumbActiveStyle = "bg-indigo-50 text-indigo-700 hover:bg-indigo-100";
