+++ frontend/app/src/screens/layout/breadcrumb-navigation/type.ts
@@ -1,0 +1,4 @@
+export type BreadcrumbItem =
+  | { type: "select"; value: string; kind: string }
+  | { type: "link"; label: string; to: string }
+  | { type: "branch"; value: string };
