+++ frontend/app/src/screens/layout/header.tsx
@@ -3,7 +3,7 @@ import { TimeFrameSelector } from "@/components/time-selector";
 import BranchSelector from "@/components/branch-selector";
 import InfrahubLogo from "@/images/infrahub-logo.svg";
 import { constructPath } from "@/utils/fetch";
-import BreadcrumbNavigation from "@/screens/layout/breadcrumb-navigation/breadcrumb";
+import BreadcrumbNavigation from "@/screens/layout/breadcrumb-navigation/breadcrumb-navigation";
 
 export default function Header() {
   return (
